# src/ViajantesTurismo.Admin.Contracts.Http/OpenApi/v1.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ViajantesTurismo.Admin.ApiService | v1",
    "version": "1.0"
  },
  "paths": {
    "/api/v1/tours": {
      "post": {
        "tags": [
          "Tours"
        ],
        "summary": "Creates a new tour.",
        "description": "Creates a new tour.",
        "operationId": "CreateTour",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTourDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTourDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The authenticated caller lacks the required permission."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "Tours"
        ],
        "summary": "Retrieves all available tours.",
        "description": "Retrieves all available tours.",
        "operationId": "GetTours",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetTourDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The authenticated caller lacks the required permission."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/tours/{id}": {
      "get": {
        "tags": [
          "Tours"
        ],
        "summary": "Retrieves a tour by its ID.",
        "description": "Retrieves a tour by its ID.",
        "operationId": "GetTourById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTourDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tours"
        ],
        "summary": "Updates an existing tour.",
        "description": "Updates an existing tour.",
        "operationId": "UpdateTour",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTourDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/bookings/tour/{tourId}": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Retrieves all bookings for a specific tour.",
        "description": "Retrieves all bookings for a specific tour.",
        "operationId": "GetBookingsByTourId",
        "parameters": [
          {
            "name": "tourId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetBookingDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/docs/errors": {
      "get": {
        "tags": [
          "Errors"
        ],
        "summary": "Retrieves generated error documentation metadata.",
        "description": "Retrieves generated documentation metadata for centralized error providers.",
        "operationId": "GetErrorDocumentation",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetErrorDocumentationDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The authenticated caller lacks the required permission."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/v1/docs/errors/{identifier}": {
      "get": {
        "tags": [
          "Errors"
        ],
        "summary": "Retrieves one generated error documentation entry.",
        "description": "Retrieves one generated centralized error documentation entry by identifier.",
        "operationId": "GetErrorDocumentationByIdentifier",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetErrorDocumentationDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          },
          "403": {
            "description": "The authenticated caller lacks the required permission."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "BookingStatusDto": {
        "type": "integer"
      },
...
      "CreateTourDto": {
        "required": [
          "identifier",
          "name",
          "startDate",
          "endDate",
          "price",
          "singleRoomSupplementPrice",
          "regularBikePrice",
          "eBikePrice",
          "currency",
          "includedServices",
          "minCustomers",
          "maxCustomers"
        ],
        "type": "object",
        "properties": {
          "identifier": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "maximum": 100000,
            "minimum": 0.01,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "singleRoomSupplementPrice": {
            "maximum": 100000,
            "minimum": 0.01,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "regularBikePrice": {
            "maximum": 100000,
            "minimum": 0.01,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "eBikePrice": {
            "maximum": 100000,
            "minimum": 0.01,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "includedServices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minCustomers": {
            "maximum": 20,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "maxCustomers": {
            "maximum": 20,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "CurrencyDto": {
        "type": "integer"
      },
...
      "DiscountTypeDto": {
        "type": "integer"
      },
...
      "GetBookingDto": {
        "required": [
          "id",
          "tourId",
          "tourIdentifier",
          "tourName",
          "customerId",
          "customerName",
          "bookingDate",
          "status",
          "paymentStatus",
          "totalPrice",
          "discountType",
          "discountAmount",
          "currency",
          "payments",
          "amountPaid",
          "remainingBalance"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tourId": {
            "type": "string",
            "format": "uuid"
          },
          "tourIdentifier": {
            "type": "string"
          },
          "tourName": {
            "type": "string"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "customerName": {
            "type": "string"
          },
          "companionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "companionName": {
            "type": [
              "null",
              "string"
            ]
          },
          "bookingDate": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatusDto"
          },
          "paymentStatus": {
            "$ref": "#/components/schemas/PaymentStatusDto"
          },
          "totalPrice": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "discountType": {
            "$ref": "#/components/schemas/DiscountTypeDto"
          },
          "discountAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "discountReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetPaymentDto"
            }
          },
          "amountPaid": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "remainingBalance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "GetErrorDocumentationDto": {
        "required": [
          "identifier",
          "documentationPath",
          "providerType",
          "memberName",
          "status",
          "httpStatusCode",
          "code",
          "detailTemplate"
        ],
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string"
          },
          "documentationPath": {
            "type": "string"
          },
          "providerType": {
            "type": "string"
          },
          "memberName": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "httpStatusCode": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "detailTemplate": {
            "type": "string"
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "GetPaymentDto": {
        "required": [
          "id",
          "bookingId",
          "amount",
          "paymentDate",
          "method",
          "recordedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "bookingId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethodDto"
          },
          "referenceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "recordedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "GetTourDto": {
        "required": [
          "id",
          "identifier",
          "name",
          "startDate",
          "endDate",
          "price",
          "singleRoomSupplementPrice",
          "regularBikePrice",
          "eBikePrice",
          "currency",
          "includedServices",
          "minCustomers",
          "maxCustomers",
          "currentCustomerCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "identifier": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "singleRoomSupplementPrice": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "regularBikePrice": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "eBikePrice": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "includedServices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minCustomers": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "maxCustomers": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currentCustomerCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "PaymentMethodDto": {
        "type": "integer"
      },
      "PaymentStatusDto": {
        "type": "integer"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateTourDto": {
        "required": [
          "identifier",
          "name",
          "startDate",
          "endDate",
          "price",
          "singleRoomSupplementPrice",
          "regularBikePrice",
          "eBikePrice",
          "currency",
          "includedServices",
          "minCustomers",
          "maxCustomers"
        ],
        "type": "object",
        "properties": {
          "identifier": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "maximum": 100000,
            "minimum": 0.01,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "singleRoomSupplementPrice": {
            "maximum": 100000,
            "minimum": 0.01,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "regularBikePrice": {
            "maximum": 100000,
            "minimum": 0.01,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "eBikePrice": {
            "maximum": 100000,
            "minimum": 0.01,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyDto"
          },
          "includedServices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "minCustomers": {
            "maximum": 20,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "maxCustomers": {
            "maximum": 20,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}